Add a shortcut to focus or unfocus a spec
This is not a *beautiful* bit of code, but I think it is quite functional.
I've often been annoyed by how I need to move my cursor to the EXACT right
location in my specs in order to focus a few tests. This shortcut allows users
to either focus, or unfocus the example under their cursor with a simple keystroke.
I chose ctrl+f, but there might be a better keybinding.

I had to add some funky code to work around an edgecase when the cursor is
inside the word "it" or "describe" or "context", but other than that, I think
this is pretty straightforward. I looked around at some of the classdumped headers
for XC classes, but couldn't find anything that would make this easier to implement.

Any suggestions for improving this code would be greatly appreciated.

I discovered that the "previous" and "next" symbol is not the next one
you would get by looking around, character by character, but xcode actually
looks around by scope.

e.g., if you have this block

it(@"sup", ^{

});
  ^  <--- and your cursor is here

Then the "previous" symbol by scope is actually "it", because it looks all
the way back past the args passed to the function.