Always show alternate file in adjacent editor
Starting in Xcode 6.3, the "open alternate" shortcut regressed.

Expected :

* User is editing Foo.m
* User activates "open alternate in adjacent" shortcut
* FooSpec.mm appears in adjacent editor

Actual :

* FooSpec.mm appears in adjacent editor, then is replaced by some other random file

I believe this occurs because we never passed along the correct document for the adjacent
editor to display. Starting in Xcode 6 there was a concept of a "fallback" document for
the IDEEditorCoordinator class, and it seems that the editor will fallback to the previous
open document if one is not provided here.