Add shortcut to add forward-declaration of classes
I do this so much that it actually made sense to have a keybinding
in Xcode to do this. Just a little bit jealous of AppCode, but I really
find it hard to switch.

This has some ... strange formatting. It ends up putting all of the
forward-declared classes just after the #imports. Would love to inject
in another object that's responsible for ... either doing the writing
to the file, or deciding what the actual string to write is.